% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age}
\usage{
calculate_age(from, to)
}
\arguments{
\item{from}{date of birth}

\item{to}{date at which age should be calculated}
}
\description{
Function to calculate age.
}
\references{
Contributed by Jim to this \href{https://stackoverflow.com/questions/3611314/calculate-ages-in-r/25450756#25450756}{StackOverflow thread}.
}
