% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_na}
\alias{check_na}
\title{Check data.frame (or equivalent) for NAs}
\usage{
check_na(dat)
}
\arguments{
\item{dat}{A data.frame, data.table, or tibble}
}
\description{
Function to count the number of NAs in each column in a data.frame.
}
