% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_bigquery.R
\name{connect_bigquery}
\alias{connect_bigquery}
\title{Establish BigQuery connection}
\usage{
connect_bigquery(
  project = Sys.getenv("BQ_PROJECT_ID"),
  billing = Sys.getenv("BQ_PROJECT_ID"),
  dataset = Sys.getenv("BQ_DATASET")
)
}
\arguments{
\item{project}{BQ project ID}

\item{billing}{BQ project ID}

\item{dataset}{BQ dataset}
}
\description{
Function to establish a BigQuery database connection.
}
