% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_redshift.R
\name{connect_redshift}
\alias{connect_redshift}
\title{Establish Redshift connection}
\usage{
connect_redshift(
  dbname = Sys.getenv("REDSHIFT_DATABASE"),
  host = Sys.getenv("REDSHIFT_HOST"),
  port = Sys.getenv("REDSHIFT_PORT"),
  user = Sys.getenv("REDSHIFT_CREDENTIAL_USERNAME"),
  password = Sys.getenv("REDSHIFT_CREDENTIAL_PASSWORD")
)
}
\arguments{
\item{dbname}{database name}

\item{host}{databse host}

\item{port}{database host}

\item{user}{database username}

\item{password}{database password}
}
\description{
Function to establish an Amazon Redshift database connection.
}
\references{
Connects to Redshift using the recommendation from
\href{https://blog.datascienceheroes.com/a-comprehensive-guide-to-connect-r-to-amazon-redshift/}{this guide}.
Importantly, it uses RPostgres, as opposed to RJDBC.
}
