% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{An R object. Currently supports numeric, character, and factor vectors.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Function to compute the mode. Supports numeric, character,
and factor variables.
}
\references{
This solution was written by
\href{https://stackoverflow.com/users/169947/ken-williams}{Ken Williams} and
adapted by \href{https://stackoverflow.com/users/903061/gregor-thomas}{Gregor Thomas}.
It comes from \href{https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode}{this}
post on Stack Overflow.
}
