% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_scientific.R
\name{no_scientific}
\alias{no_scientific}
\title{Disable scientific notation}
\usage{
no_scientific()
}
\description{
Function to disable scientific notation.
}
\references{
Uses the following Stack Overflow
\href{https://stackoverflow.com/questions/25946047/how-to-prevent-scientific-notation-in-r}{solution}.
}
