% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saves3folder.R
\name{saves3folder}
\alias{saves3folder}
\title{Downloads the contents of an S3 bucket or folder to local}
\usage{
saves3folder(bucket, path, prefix = NULL)
}
\arguments{
\item{bucket}{S3 bucket name, excluding "s3://"}

\item{path}{path to local directory in which to copy the folder}

\item{prefix}{optional. Path from the bucket to the directory, ending in a slash}
}
\description{
Essentially applies \code{aws.s3::save_object()} to all objects inside of
an S3 bucket or folder.
}
