% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_case.R
\name{snake_case}
\alias{snake_case}
\title{Snake case}
\usage{
snake_case(input)
}
\arguments{
\item{input}{A string or vector of strings}
}
\description{
Function to convert string or vector of strings to snake case
(e.g., snake_case).
}
