% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{Cross Tabulation and Table Creation}
\usage{
table2(
  ...,
  exclude = if (useNA == "no") c(NA, NaN),
  useNA = "always",
  dnn = list.names(...),
  deparse.level = 1
)
}
\description{
Equivalent of base::table(), except where useNA = "always" by default. See
documentation for base::table() for more information.
}
