% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_custom}
\alias{theme_custom}
\title{Apply custom ggplot2 theme}
\usage{
theme_custom(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base line size}

\item{base_rect_size}{base rectangle size}
}
\description{
Function containing a custom ggplot2 theme, modified from \code{theme_bw()}.
}
