% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{Cross Tabulation and Table Creation}
\usage{
table2(..., useNA = "always")
}
\arguments{
\item{...}{one or more objects which can be interpreted as factors}

\item{useNA}{by default set to 'always'}
}
\description{
Equivalent of base::table(), except where useNA = "always" by default. See
documentation for base::table() for more information.
}
